Repo and Service methods done writing

+++ src/main/java/com/example/service/MessageService.java
@@ -22,13 +22,21 @@ public class MessageService {
     }
 
     /**
-     * Adds a message. Assumes it's valid by following these parameters:
+     * Adds a message. Checks for follwoing validality:
      * 1. messageText is not empty nor greater than 255 characters
      * 2. postedBy refers to a real existing account
      * @return the newly minted message 
      */
-    public Message postMessage(Message message){
-        return messageRepository.save(message);
+    public Message postMessage(int postedBy, String messageBody, long postedOn){
+        if (messageBody.length() > 0 && messageBody.length() <= 255){
+            // message text is valid; is postedBy valid?
+            Optional<Account> account = accountRepository.findById((long)postedBy);
+            if (account.isPresent()){
+                // valid postedBy
+                return messageRepository.save( new Message(postedBy, messageBody, postedOn) );
+            }
+        }
+        return null;
     }
 
     /**
@@ -54,18 +62,33 @@ public class MessageService {
     }
 
     /**
-     * Deletes message by its ID. Does not check if it actually exists.
+     * Deletes message by its ID.
+     * @return did it delete a message?
      */
-    public void deleteMessageByID(long id){
-        messageRepository.deleteById(id);
+    public boolean deleteMessageByID(long id){
+        if (messageRepository.findById(id).isPresent()){
+            messageRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 
-    public Message updateMessage(Message message){
-        if (getMessageByID(message.getMessageId()) != null){
 
+    public Message updateMessage(long id, String newMessageText){
+        if (newMessageText.length() > 0 && newMessageText.length() <= 255)
+            return null;
+        Optional<Message> OptMessage = messageRepository.findById(id);
+        if (OptMessage.isPresent()){
+            Message message = OptMessage.get();
+            message.setMessageText(newMessageText);
+            return messageRepository.save(message);
         }
         else{
             return null;
         }
+    }
+
+    public List<Message> getAllMessagesByAccountID(int accountId){
+        return messageRepository.findMessagesByPostedBy(accountId);
     }
 }
